add addPeople endpint that uses POST to add new itmes into the database
note that this endpoint inserts a bulk - many ites when you send an
array.
to insert single itme just use an array with only one item in it.
a single item without an array brackets dont work
but i can implement it by looking into the received data
and create a serializer for a single object (many = False)